backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.6.0
django-filter==24.3
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
httpx==0.27.2
beautifulsoup4==4.12.3
anthropic==0.40.0
googlemaps==4.10.0
Pillow==11.1.0
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
